Add ModelTrait with `has` and `get` methods
